ide: PDB debugger U++ pretty displaying
git-svn-id: svn://ultimatepp.org/upp/trunk@13706 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/ide/Debuggers/Code.cpp
@@ -193,7 +193,7 @@ bool Pdb::ConditionalPass()
 					Array<Frame> frame = Backtrace(ctx, true);
 					current_frame = &frame[0];
 					CParser p(exp);
-					if(!GetInt(Exp(p))) {
+					if(!GetInt64(Exp(p))) {
 						current_frame = NULL;
 						return true;
 					}
